refactor(tensorboard): move TensorBoardWriter to gigl/utils
Relocate from gigl/src/common/utils/tensorboard.py (internal pipeline
utilities) to gigl/utils/tensorboard_writer.py (general-purpose user
utilities). Update the test file path, the in-test patch paths, and
all four example training scripts to import from the new location.
