+++ tests/unit/utils/tensorboard_writer_test.py
@@ -1,4 +1,4 @@
-"""Unit tests for gigl.src.common.utils.tensorboard."""
+"""Unit tests for gigl.utils.tensorboard_writer."""
 
 import os
 from unittest.mock import Mock, patch
@@ -6,7 +6,7 @@ from unittest.mock import Mock, patch
 from absl.testing import absltest
 
 from gigl.common import UriFactory
-from gigl.src.common.utils.tensorboard import TensorBoardWriter
+from gigl.utils.tensorboard_writer import TensorBoardWriter
 from tests.test_assets.test_case import TestCase
 
 
@@ -16,7 +16,7 @@ class TestTensorBoardWriter(TestCase):
     def test_from_uri_returns_noop_when_disabled(self) -> None:
         configured_uri = UriFactory.create_uri("gs://config/logs/")
         with patch(
-            "gigl.src.common.utils.tensorboard.tf.summary.create_file_writer"
+            "gigl.utils.tensorboard_writer.tf.summary.create_file_writer"
         ) as mock_create_file_writer:
             writer = TensorBoardWriter.from_uri(configured_uri, enabled=False)
             writer.log({"Loss/train": 1.0}, step=0)
@@ -32,7 +32,7 @@ class TestTensorBoardWriter(TestCase):
             clear=False,
         ):
             with patch(
-                "gigl.src.common.utils.tensorboard.tf.summary.create_file_writer"
+                "gigl.utils.tensorboard_writer.tf.summary.create_file_writer"
             ) as mock_create_file_writer:
                 TensorBoardWriter.from_uri(configured_uri)
 
@@ -42,7 +42,7 @@ class TestTensorBoardWriter(TestCase):
         configured_uri = UriFactory.create_uri("gs://config/logs/")
         with patch.dict(os.environ, {}, clear=True):
             with patch(
-                "gigl.src.common.utils.tensorboard.tf.summary.create_file_writer"
+                "gigl.utils.tensorboard_writer.tf.summary.create_file_writer"
             ) as mock_create_file_writer:
                 TensorBoardWriter.from_uri(configured_uri)
 
@@ -51,7 +51,7 @@ class TestTensorBoardWriter(TestCase):
     def test_from_uri_returns_noop_when_no_uri_anywhere(self) -> None:
         with patch.dict(os.environ, {}, clear=True):
             with patch(
-                "gigl.src.common.utils.tensorboard.tf.summary.create_file_writer"
+                "gigl.utils.tensorboard_writer.tf.summary.create_file_writer"
             ) as mock_create_file_writer:
                 writer = TensorBoardWriter.from_uri(configured_uri=None)
                 writer.log({"Loss/train": 1.0}, step=0)
@@ -59,7 +59,7 @@ class TestTensorBoardWriter(TestCase):
 
         mock_create_file_writer.assert_not_called()
 
-    @patch("gigl.src.common.utils.tensorboard.tf.summary.scalar")
+    @patch("gigl.utils.tensorboard_writer.tf.summary.scalar")
     def test_log_writes_each_metric_at_step_and_flushes(
         self, mock_summary_scalar
     ) -> None:
@@ -67,7 +67,7 @@ class TestTensorBoardWriter(TestCase):
         underlying_writer.as_default.return_value.__enter__ = Mock(return_value=None)
         underlying_writer.as_default.return_value.__exit__ = Mock(return_value=None)
         with patch(
-            "gigl.src.common.utils.tensorboard.tf.summary.create_file_writer",
+            "gigl.utils.tensorboard_writer.tf.summary.create_file_writer",
             return_value=underlying_writer,
         ):
             writer = TensorBoardWriter(log_dir="gs://logs/")
@@ -78,10 +78,10 @@ class TestTensorBoardWriter(TestCase):
         mock_summary_scalar.assert_any_call("Loss/val", 2.0, step=10)
         underlying_writer.flush.assert_called_once()
 
-    @patch("gigl.src.common.utils.tensorboard.tf.summary.scalar")
+    @patch("gigl.utils.tensorboard_writer.tf.summary.scalar")
     def test_log_is_noop_when_writer_disabled(self, mock_summary_scalar) -> None:
         with patch(
-            "gigl.src.common.utils.tensorboard.tf.summary.create_file_writer"
+            "gigl.utils.tensorboard_writer.tf.summary.create_file_writer"
         ) as mock_create_file_writer:
             writer = TensorBoardWriter(log_dir=None)
             writer.log({"Loss/train": 1.0}, step=0)
@@ -92,7 +92,7 @@ class TestTensorBoardWriter(TestCase):
     def test_context_manager_closes_writer(self) -> None:
         underlying_writer = Mock()
         with patch(
-            "gigl.src.common.utils.tensorboard.tf.summary.create_file_writer",
+            "gigl.utils.tensorboard_writer.tf.summary.create_file_writer",
             return_value=underlying_writer,
         ):
             with TensorBoardWriter(log_dir="gs://logs/"):
@@ -103,7 +103,7 @@ class TestTensorBoardWriter(TestCase):
     def test_close_is_idempotent(self) -> None:
         underlying_writer = Mock()
         with patch(
-            "gigl.src.common.utils.tensorboard.tf.summary.create_file_writer",
+            "gigl.utils.tensorboard_writer.tf.summary.create_file_writer",
             return_value=underlying_writer,
         ):
             writer = TensorBoardWriter(log_dir="gs://logs/")
